degats varie selon le type de l'allie/l'enemy ( systeme de resistance)

+++ Project/Entities/EntityCapacities.cs
@@ -1,7 +1,7 @@
 ﻿public class EntitiesCapacities
 {
     public string _name { get; set; }
-    public int _damage { get; set; }
+    public float _damage { get; set; }
     public int _stamina { get; set; }
     public int _speed { get; set; }
     public float _resistance { get; set; }
